#162 moved the functions from timer.js to util.js
