+++ lib/index.js
@@ -16,7 +16,6 @@ const MetadataTypeInfo = require('./MetadataTypeInfo');
 const MetadataTypeDefinitions = require('./MetadataTypeDefinitions');
 const Retriever = require('./Retriever');
 const cache = require('./util/cache');
-const Timer = require('./util/timer');
 
 /**
  * main class
@@ -54,12 +53,12 @@ class Mcdev {
      * @returns {Promise.<TYPE.DeltaPkgItem[]>} list of changed items
      */
     static async createDeltaPkg(argv) {
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         Util.logger.info('Create Delta Package ::');
         Mcdev.setSkipInteraction(argv.skipInteraction);
         const properties = await config.getProperties();
         if (!(await config.checkProperties(properties))) {
-            Timer.stopTimer(startTime, 'createDeltaPkg');
+            Util.stopTimer(startTime, 'createDeltaPkg');
             return null;
         }
 
@@ -68,7 +67,7 @@ class Mcdev {
               DevOps.getDeltaList(properties, argv.range, true, argv.filter)
             : // If no custom filter was provided, use deployment marketLists & templating
               DevOps.buildDeltaDefinitions(properties, argv.range, argv.skipInteraction);
-        Timer.stopTimer(startTime, 'createDeltaPkg');
+        Util.stopTimer(startTime, 'createDeltaPkg');
         return response;
     }
 
@@ -117,11 +116,11 @@ class Mcdev {
      */
     static async retrieve(businessUnit, selectedTypesArr, keys, changelogOnly) {
         Util.logger.info('mcdev:: Retrieve');
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         const properties = await config.getProperties();
         if (!(await config.checkProperties(properties))) {
             // return null here to avoid seeing 2 error messages for the same issue
-            Timer.stopTimer(startTime, 'retrieve');
+            Util.stopTimer(startTime, 'retrieve');
             return null;
         }
 
@@ -129,7 +128,7 @@ class Mcdev {
         if (selectedTypesArr) {
             for (const selectedType of selectedTypesArr) {
                 if (!Util._isValidType(selectedType)) {
-                    Timer.stopTimer(startTime, 'retrieve');
+                    Util.stopTimer(startTime, 'retrieve');
                     return;
                 }
             }
@@ -168,7 +167,7 @@ class Mcdev {
                     cred = buObject.credential;
                     bu = buObject.businessUnit;
                 } else {
-                    Timer.stopTimer(startTime, 'retrieve');
+                    Util.stopTimer(startTime, 'retrieve');
                     return;
                 }
             }
@@ -192,7 +191,7 @@ class Mcdev {
                     changelogOnly
                 );
                 if (changelogOnly) {
-                    Timer.stopTimer(startTime, 'retrieve');
+                    Util.stopTimer(startTime, 'retrieve');
                     return retrieveChangelog;
                 }
                 Util.logger.info(`:: Done\n`);
@@ -288,9 +287,9 @@ class Mcdev {
      * @returns {Promise.<void>} -
      */
     static async deploy(businessUnit, selectedTypesArr, keyArr, fromRetrieve = false) {
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         const response = Deployer.deploy(businessUnit, selectedTypesArr, keyArr, fromRetrieve);
-        Timer.stopTimer(startTime, 'deploy');
+        Util.stopTimer(startTime, 'deploy');
         return response;
     }
 
@@ -479,14 +478,14 @@ class Mcdev {
      */
     static async retrieveAsTemplate(businessUnit, selectedType, name, market) {
         Util.logger.info('mcdev:: Retrieve as Template');
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         const properties = await config.getProperties();
         if (!(await config.checkProperties(properties))) {
-            Timer.stopTimer(startTime, 'retrieveAsTemplate');
+            Util.stopTimer(startTime, 'retrieveAsTemplate');
             return null;
         }
         if (!Util._isValidType(selectedType)) {
-            Timer.stopTimer(startTime, 'retrieveAsTemplate');
+            Util.stopTimer(startTime, 'retrieveAsTemplate');
             return;
         }
         const [type, subType] = selectedType ? selectedType.split('-') : [];
@@ -512,7 +511,7 @@ class Mcdev {
                     name,
                     properties.markets[market]
                 );
-                Timer.stopTimer(startTime, 'retrieveAsTemplate');
+                Util.stopTimer(startTime, 'retrieveAsTemplate');
                 return response;
             }
         }
@@ -528,10 +527,10 @@ class Mcdev {
      * @returns {Promise.<TYPE.MultiMetadataTypeList>} -
      */
     static async buildTemplate(businessUnit, selectedType, keyArr, market) {
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         Util.logger.info('mcdev:: Build Template from retrieved files');
         const response = Builder.buildTemplate(businessUnit, selectedType, keyArr, market);
-        Timer.stopTimer(startTime, 'buildTemplate');
+        Util.stopTimer(startTime, 'buildTemplate');
         return response;
     }
     /**
@@ -544,10 +543,10 @@ class Mcdev {
      * @returns {Promise.<void>} -
      */
     static async buildDefinition(businessUnit, selectedType, name, market) {
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         Util.logger.info('mcdev:: Build Definition from Template');
         const response = Builder.buildDefinition(businessUnit, selectedType, name, market);
-        Timer.stopTimer(startTime, 'buildDefinition');
+        Util.stopTimer(startTime, 'buildDefinition');
         return response;
     }
 
@@ -561,9 +560,9 @@ class Mcdev {
      */
     static async buildDefinitionBulk(listName, type, name) {
         Util.logger.info('mcdev:: Build Definition from Template Bulk');
-        const startTime = Timer.startTimer();
+        const startTime = Util.startTimer();
         const response = Builder.buildDefinitionBulk(listName, type, name);
-        Timer.stopTimer(startTime, 'buildDefinitionBulk');
+        Util.stopTimer(startTime, 'buildDefinitionBulk');
         return response;
     }
     /**
