+++ lib/util/util.js
@@ -125,6 +125,25 @@ const Util = {
         }
     },
     /**
+     * calculates the runtime for a function
+     *
+     * @returns {number} - actual time in ms
+     */
+    startTimer() {
+        return performance.now();
+    },
+    /**
+     * stops the timer and print the ms needed for the command
+     *
+     * @param {number} initTime value in ms from startTimer()
+     * @param {string} command command that is being currently timed
+     * @returns {void}
+     */
+    stopTimer(initTime, command) {
+        const totalTime = performance.now() - initTime;
+        Util.logger.info(`:::  The ${command} command runtime was ${totalTime.toFixed(2)} ms`);
+    },
+    /**
      * used to ensure the program tells surrounding software that an unrecoverable error occured
      *
      * @returns {void}
@@ -419,7 +438,7 @@ const Util = {
             Util.loggerTransports.console.level = 'debug';
             Util.logger.debug('CLI logger set to: debug');
         }
-    },
+    }
 };
 /**
  * wrapper around our standard winston logging to console and logfile
@@ -436,7 +455,7 @@ function createNewLoggerTransport() {
                 winston.format.timestamp({ format: 'HH:mm:ss' }),
                 winston.format.simple(),
                 winston.format.printf((info) => `${info.timestamp} ${info.level}: ${info.message}`)
-            ),
+            )
         }),
         file: new winston.transports.File({
             // Write logs to logfile
@@ -447,8 +466,8 @@ function createNewLoggerTransport() {
                 winston.format.timestamp({ format: 'HH:mm:ss.SSS' }),
                 winston.format.simple(),
                 winston.format.printf((info) => `${info.timestamp} ${info.level}: ${info.message}`)
-            ),
-        }),
+            )
+        })
     };
 }
 /**
@@ -460,7 +479,7 @@ function startLogger() {
     Util.loggerTransports = createNewLoggerTransport();
     const myWinston = winston.createLogger({
         levels: winston.config.npm.levels,
-        transports: [Util.loggerTransports.console, Util.loggerTransports.file],
+        transports: [Util.loggerTransports.console, Util.loggerTransports.file]
     });
     const winstonError = myWinston.error;
     const winstonExtension = {
@@ -501,7 +520,7 @@ function startLogger() {
         error: function (msg) {
             winstonError(msg);
             Util.signalFatalError();
-        },
+        }
     };
     Util.logger = Object.assign(myWinston, winstonExtension);
